+++ sdk/redis/arm-rediscache/src/operations/redis.ts
@@ -373,86 +373,16 @@ export class RedisImpl implements Redis {
    * @param parameters Parameters supplied to the Update Redis operation.
    * @param options The options parameters.
    */
-  async beginUpdate(
-    resourceGroupName: string,
-    name: string,
-    parameters: RedisUpdateParameters,
-    options?: RedisUpdateOptionalParams
-  ): Promise<
-    PollerLike<PollOperationState<RedisUpdateResponse>, RedisUpdateResponse>
-  > {
-    const directSendOperation = async (
-      args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
-    ): Promise<RedisUpdateResponse> => {
-      return this.client.sendOperationRequest(args, spec);
-    };
-    const sendOperation = async (
-      args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
-    ) => {
-      let currentRawResponse:
-        | coreClient.FullOperationResponse
-        | undefined = undefined;
-      const providedCallback = args.options?.onResponse;
-      const callback: coreClient.RawResponseCallback = (
-        rawResponse: coreClient.FullOperationResponse,
-        flatResponse: unknown
-      ) => {
-        currentRawResponse = rawResponse;
-        providedCallback?.(rawResponse, flatResponse);
-      };
-      const updatedArgs = {
-        ...args,
-        options: {
-          ...args.options,
-          onResponse: callback
-        }
-      };
-      const flatResponse = await directSendOperation(updatedArgs, spec);
-      return {
-        flatResponse,
-        rawResponse: {
-          statusCode: currentRawResponse!.status,
-          body: currentRawResponse!.parsedBody,
-          headers: currentRawResponse!.headers.toJSON()
-        }
-      };
-    };
-
-    const lro = new LroImpl(
-      sendOperation,
-      { resourceGroupName, name, parameters, options },
-      updateOperationSpec
-    );
-    const poller = new LroEngine(lro, {
-      resumeFrom: options?.resumeFrom,
-      intervalInMs: options?.updateIntervalInMs
-    });
-    await poller.poll();
-    return poller;
-  }
-
-  /**
-   * Update an existing Redis cache.
-   * @param resourceGroupName The name of the resource group.
-   * @param name The name of the Redis cache.
-   * @param parameters Parameters supplied to the Update Redis operation.
-   * @param options The options parameters.
-   */
-  async beginUpdateAndWait(
+  update(
     resourceGroupName: string,
     name: string,
     parameters: RedisUpdateParameters,
     options?: RedisUpdateOptionalParams
   ): Promise<RedisUpdateResponse> {
-    const poller = await this.beginUpdate(
-      resourceGroupName,
-      name,
-      parameters,
-      options
+    return this.client.sendOperationRequest(
+      { resourceGroupName, name, parameters, options },
+      updateOperationSpec
     );
-    return poller.pollUntilDone();
   }
 
   /**
@@ -945,15 +875,6 @@ const updateOperationSpec: coreClient.OperationSpec = {
   httpMethod: "PATCH",
   responses: {
     200: {
-      bodyMapper: Mappers.RedisResource
-    },
-    201: {
-      bodyMapper: Mappers.RedisResource
-    },
-    202: {
-      bodyMapper: Mappers.RedisResource
-    },
-    204: {
       bodyMapper: Mappers.RedisResource
     },
     default: {